Apply defaults more elegantly
