+++ packages/use-intl/src/core/defaults.tsx
@@ -1,0 +1,17 @@
+import IntlError from './IntlError';
+
+/**
+ * Contains defaults that are used for all entry points into the core.
+ */
+
+export function defaultGetMessageFallback(props: {
+  error: IntlError;
+  key: string;
+  namespace?: string;
+}) {
+  return [props.namespace, props.key].filter((part) => part != null).join('.');
+}
+
+export function defaultOnError(error: IntlError) {
+  console.error(error);
+}
